svn path=/branches/ros-csrss/; revision=58607

+++ dll/win32/ole32/storage32.c
@@ -1767,7 +1767,7 @@ static HRESULT WINAPI StorageBaseImpl_CopyTo(
   StorageBaseImpl *This = impl_from_IStorage(iface);
 
   BOOL         skip_storage = FALSE, skip_stream = FALSE;
-  int          i;
+  DWORD        i;
 
   TRACE("(%p, %d, %p, %p, %p)\n",
 	iface, ciidExclude, rgiidExclude,
@@ -2822,7 +2822,7 @@ static HRESULT StorageImpl_Construct(
   {
     ULONG current_block = This->extBigBlockDepotStart;
     ULONG cache_size = This->extBigBlockDepotCount * 2;
-    int i;
+    ULONG i;
 
     This->extBigBlockDepotLocations = HeapAlloc(GetProcessHeap(), 0, sizeof(ULONG) * cache_size);
     if (!This->extBigBlockDepotLocations)
